requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Database & ORM
sqlalchemy==2.0.36
greenlet==3.1.1
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config & validation
pydantic==2.10.3
pydantic-settings==2.6.1
